devtools/client/shared/components/reps/test/mochitest/test_reps_null.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test Null rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - Null</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript;version=1.8"></script>
<script type="application/javascript;version=1.8">
window.onload = Task.async(function* () {
  try {
    const { REPS } = browserRequire("devtools/client/shared/components/reps/load-reps");
    let { Rep, Null } = REPS;

    let gripStub = {
      "type": "null"
    };

    // Test that correct rep is chosen
    const renderedRep = shallowRenderComponent(Rep, { object: gripStub });
    is(renderedRep.type, Null.rep, `Rep correctly selects ${Null.rep.displayName}`);

    // Test rendering
    const renderedComponent = renderComponent(Null.rep, { object: gripStub });
    is(renderedComponent.textContent, "null", "Null rep has expected text content");
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
});
</script>
</pre>
</body>
</html>
